Unit tests + Lint comments

+++ gocache.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 )
 
+// GoCacheBackend acts as a wrapper around a Backend interface
 type GoCacheBackend struct {
 	Backend
 	Connection *cache.Cache
@@ -15,6 +16,7 @@ type GoCacheBackend struct {
 	Source     string
 }
 
+// NewGoCache initialises a new database using the BadgerDB driver
 func NewGoCache(source string, memory bool) (Backend, error) {
 	if source == "" {
 		source = "data"
@@ -37,13 +39,14 @@ func NewGoCache(source string, memory bool) (Backend, error) {
 	return &database, nil
 }
 
+// Count returns the total number of records inside of the provided bucket
 func (database *GoCacheBackend) Count(bucket string) (int, error) {
 	db := database.Connection
 	counter := 0
 
 	data := db.Items()
 
-	for key, _ := range data {
+	for key := range data {
 		if strings.HasPrefix(key, bucket+"_") {
 			counter++
 		}
@@ -52,6 +55,7 @@ func (database *GoCacheBackend) Count(bucket string) (int, error) {
 	return counter, nil
 }
 
+// Create inserts a record into the backend
 func (database *GoCacheBackend) Create(bucket string, key string, model interface{}) error {
 	db := database.Connection
 
@@ -73,8 +77,14 @@ func (database *GoCacheBackend) Create(bucket string, key string, model interfac
 	return nil
 }
 
+// Delete removes a record from the backend
 func (database *GoCacheBackend) Delete(bucket string, key string) error {
 	db := database.Connection
+
+	_, found := db.Get(bucket + "_" + key)
+	if !found {
+		return errors.New("key could not be found")
+	}
 
 	db.Delete(bucket + "_" + key)
 
@@ -87,6 +97,7 @@ func (database *GoCacheBackend) Delete(bucket string, key string) error {
 	return nil
 }
 
+// Get returns all records inside of the provided bucket
 func (database *GoCacheBackend) Get(bucket string, model interface{}) (*map[string]interface{}, error) {
 	db := database.Connection
 	results := make(map[string]interface{})
@@ -106,6 +117,7 @@ func (database *GoCacheBackend) Get(bucket string, model interface{}) (*map[stri
 	return &results, nil
 }
 
+// Read returns a single struct from the provided bucket, using the provided key
 func (database *GoCacheBackend) Read(bucket string, key string, model interface{}) error {
 	db := database.Connection
 
@@ -117,10 +129,16 @@ func (database *GoCacheBackend) Read(bucket string, key string, model interface{
 	return json.Unmarshal(data.([]byte), &model)
 }
 
+// Update modifies an existing record from the backend, inside of the provided bucket, using the provided key
 func (database *GoCacheBackend) Update(bucket string, key string, model interface{}) error {
 	db := database.Connection
 
-	if err := db.Replace(bucket+"_"+key, model, cache.NoExpiration); err != nil {
+	data, err := json.Marshal(&model)
+	if err != nil {
+		return err
+	}
+
+	if err := db.Replace(bucket+"_"+key, data, cache.NoExpiration); err != nil {
 		return err
 	}
